Initial commit of fluentd example filter
The purpose of this project was to bootstrap implementing a new
filter.  The example project has some comments in it as well.
For the most part, you'll just want to change the
`def filter(tag, time, record)` method in the library file.  The
test files also have some examples of how you can test whether
or not your configs are getting properly set, as well as testing
whether or not the records that were emitted were properly filtered.
Finally, there's an example benchmark test in case you want to
determine the speed of your filter.